fix : move confAccessTokenIsSaved method
